+++ bom/pom.xml
@@ -2,7 +2,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <groupId>io.github.marcelorubim</groupId>
+        <groupId>br.com.labbs</groupId>
         <artifactId>quarkus-monitor-parent</artifactId>
       <version>0.1.4-SNAPSHOT</version>
     </parent>
