+++ deployment/pom.xml
@@ -59,7 +59,7 @@
     </dependency>
     <dependency>
       <artifactId>quarkus-monitor</artifactId>
-      <groupId>io.github.marcelorubim</groupId>
+      <groupId>br.com.labbs</groupId>
       <version>${project.version}</version>
     </dependency>
     <dependency>
@@ -82,7 +82,7 @@
 
   <parent>
     <artifactId>quarkus-monitor-parent</artifactId>
-    <groupId>io.github.marcelorubim</groupId>
+    <groupId>br.com.labbs</groupId>
     <relativePath>../</relativePath>
     <version>0.1.4-SNAPSHOT</version>
   </parent>
