+++ deployment/src/test/java/br/com/rubim/test/fake/resources/DependencyResource.java
@@ -1,4 +1,4 @@
-package br.com.rubim.test.fake.resources;
+package br.com.labbs.quarkusmonitor.test.fake.resources;
 
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
