+++ deployment/src/test/java/br/com/rubim/test/metrics/DependencyUpMetricsTest.java
@@ -1,15 +1,15 @@
-package br.com.rubim.test.metrics;
+package br.com.labbs.quarkusmonitor.test.metrics;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTimeout;
 import static org.junit.jupiter.api.Assertions.fail;
 
-import br.com.rubim.runtime.MonitorMetrics;
-import br.com.rubim.runtime.dependency.DependencyState;
-import br.com.rubim.test.fake.filters.MetricsFilterForError;
-import br.com.rubim.test.fake.resources.DependencyResource;
-import br.com.rubim.test.fake.resources.DependencyRestClient;
+import br.com.labbs.quarkusmonitor.runtime.MonitorMetrics;
+import br.com.labbs.quarkusmonitor.runtime.dependency.DependencyState;
+import br.com.labbs.quarkusmonitor.test.fake.filters.MetricsFilterForError;
+import br.com.labbs.quarkusmonitor.test.fake.resources.DependencyResource;
+import br.com.labbs.quarkusmonitor.test.fake.resources.DependencyRestClient;
 import io.micrometer.core.instrument.Gauge;
 import io.micrometer.core.instrument.Metrics;
 import io.quarkus.test.QuarkusUnitTest;
@@ -35,7 +35,7 @@ public class DependencyUpMetricsTest {
               DependencyRestClient.class)
           .addAsResource(
               new StringAsset("quarkus.b5.monitor.enable-http-response-size=false\n" +
-                  "br.com.rubim.test.fake.resources.DependencyRestClient/mp-rest/url=${test.url}"),
+                  "br.com.labbs.quarkusmonitor.test.fake.resources.DependencyRestClient/mp-rest/url=${test.url}"),
               "application.properties"));
 
   @RestClient
