+++ pom.xml
@@ -164,7 +164,7 @@
       <url>https://oss.sonatype.org/content/repositories/snapshots</url>
     </snapshotRepository>
   </distributionManagement>
-  <groupId>io.github.marcelorubim</groupId>
+  <groupId>br.com.labbs</groupId>
   <licenses>
     <license>
       <distribution>repo</distribution>
@@ -211,13 +211,13 @@
   </properties>
 
   <scm>
-    <connection>scm:git:https://github.com/marcelorubim/quarkus-monitor/</connection>
-    <developerConnection>scm:git:https://github.com/marcelorubim/quarkus-monitor/
+    <connection>scm:git:https://github.com/labbsr0x/quarkus-monitor/</connection>
+    <developerConnection>scm:git:https://github.com/labbsr0x/quarkus-monitor/
     </developerConnection>
     <tag>0.1.0-SNAPSHOT</tag>
-    <url>https://github.com/marcelorubim/quarkus-monitor/</url>
+    <url>https://github.com/labbsr0x/quarkus-monitor/</url>
   </scm>
-  <url>https://github.com/marcelorubim/quarkus-monitor/</url>
+  <url>https://github.com/labbsr0x/quarkus-monitor/</url>
   <version>0.1.4-SNAPSHOT</version>
 
 </project>
