+++ runtime/src/main/java/br/com/labbs/quarkusmonitor/runtime/filters/MetricsServiceFilter.java
@@ -1,9 +1,9 @@
-package br.com.rubim.runtime.filters;
+package br.com.labbs.quarkusmonitor.runtime.filters;
 
-import br.com.rubim.runtime.MonitorMetrics;
-import br.com.rubim.runtime.core.Metrics;
-import br.com.rubim.runtime.util.FilterUtils;
-import br.com.rubim.runtime.util.TagsUtil;
+import br.com.labbs.quarkusmonitor.runtime.MonitorMetrics;
+import br.com.labbs.quarkusmonitor.runtime.core.Metrics;
+import br.com.labbs.quarkusmonitor.runtime.util.FilterUtils;
+import br.com.labbs.quarkusmonitor.runtime.util.TagsUtil;
 import java.io.IOException;
 import java.time.Instant;
 import java.util.Optional;
