+++ runtime/src/main/java/br/com/labbs/quarkusmonitor/runtime/util/TagsUtil.java
@@ -1,4 +1,4 @@
-package br.com.rubim.runtime.util;
+package br.com.labbs.quarkusmonitor.runtime.util;
 
 import java.util.Optional;
 import javax.ws.rs.client.ClientRequestContext;
